# openapi/openapi.generated.yaml
openapi: 3.0.0
info:
  title: Generated routes
  version: 1.0.0
paths:
  "/api/admin/activity":
    get:
          summary: Generated stub (GET)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
...
  "/api/admin/identity-changes":
    get:
          summary: Generated stub (GET)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
...
  "/api/admin/message":
    post:
          summary: Generated stub (POST)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
...
  "/api/auth/forgot-password":
    post:
          summary: Generated stub (POST)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
...
  "/api/identity-change":
    get:
          summary: Generated stub (GET)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
    post:
          summary: Generated stub (POST)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
...
  "/api/invite/{token}":
    get:
          summary: Generated stub (GET)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
    post:
          summary: Generated stub (POST)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
...
  "/api/members/{userId}":
    get:
          summary: Generated stub (GET)
          responses:
            "200":
              description: OK
            "4XX":
              description: Client error
            "5XX":
              description: Server error
...